Refactoring: move kakasi.py to __init__.py
