+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = mypy, clean, check, py36, py37, py38, py39, pypy3, docs, report
+envlist = mypy, clean, check, py36, py37, py38, py39, pypy3, docs
 
 [testenv]
 extras = test
@@ -26,7 +26,7 @@ extras = check
 commands =
     check-manifest {toxinidir}
     flake8 src tests setup.py
-    isort --verbose --check-only --diff --recursive src tests setup.py
+    isort --check-only --diff src tests setup.py
 
 [testenv:clean]
 deps = coverage[toml]>=5.2
@@ -73,14 +73,6 @@ extras = docs
 commands =
     sphinx-build {posargs:-E} -b html docs build/docs
     sphinx-build -b linkcheck docs build/docs
-
-[testenv:report]
-basepython = python3.8
-deps = coverage[toml]>=5.2
-skip_install = true
-commands =
-    coverage report
-    coverage html -d build/htmlcov
 
 [flake8]
 extend-ignore = E203, W503
